Moved database functionality
Moved from Search Activity to Database class
