+++ code/app/src/main/java/com/example/booknet/MockDatabase.java
@@ -152,8 +152,7 @@ public class MockDatabase extends DatabaseManager {
     }
 
     @Override
-    public ArrayList<BookListing> readAllBookListings() {
-        return bookListings;
+    public void readAllBookListings(BookSearchActivity activity) {
     }
 
     @Override
